feat(package.json): add dependencies for axios, redux, react-redux, and react-router-dom refactor(App.js): update import paths for Greeting component and store refactor(Greeting.js): update import path for getMessage function and change API endpoint to use localhost instead of IP address refactor(configureStore.js): update import path for greetingsReducer
