+++ src/App.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { Provider } from "react-redux";
-import Greeting from "./Greeting";
-import store from "../redux/configureStore";
+import Greeting from "./components/Greeting";
+import store from "./redux/configureStore";
 
 const App = () => {
   return (
