+++ src/components/Greeting.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { getMessage } from "../redux/greetingsReducer";
+import { getMessage } from "../redux/greetingReducer";
 import axios from "axios";
 
 const Greeting = () => {
@@ -8,7 +8,7 @@ const Greeting = () => {
   const message = useSelector((state) => state.greeting);
 
   async function fetchMessage() {
-    await axios.get("api/messages").then((response) => {
+    await axios.get("http://127.0.0.1:3000/api/messages").then((response) => {
       dispatch(getMessage(response.data.greeting));
     });
   }
