+++ src/redux/configureStore.js
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import greetingsReducer from "./greetingsReducer";
+import greetingsReducer from "./greetingReducer";
 
 const store = configureStore({
   reducer: {
